+++ datapipe/sql_util.py
@@ -8,29 +8,6 @@ from sqlalchemy.sql.expression import and_, or_
 from datapipe.run_config import RunConfig
 from datapipe.types import IndexDF, LabelDict
 
-
-def sql_apply_filters_idx_to_subquery(
-    sql: Any,
-    keys: List[str],
-    filters_idx: Optional[pd.DataFrame],
-) -> Any:
-    if filters_idx is None:
-        return sql
-
-    applicable_filter_keys = [i for i in filters_idx.columns if i in keys]
-    if len(applicable_filter_keys) > 0:
-        sql = sql.where(
-            tuple_(*[column(i) for i in applicable_filter_keys]).in_(
-                [
-                    tuple_(*[r[k] for k in applicable_filter_keys])
-                    for r in filters_idx.to_dict(orient="records")
-                ]
-            )
-        )
-
-    return sql
-
-
 def sql_apply_idx_filter_to_table(
     sql: Any,
     table: Table,
@@ -60,12 +37,14 @@ def sql_apply_idx_filter_to_table(
 
 def sql_apply_runconfig_filters(
     sql: Any,
+    table: Table,
     keys: List[str],
     run_config: Optional[RunConfig] = None,
 ) -> Any:
     if run_config is not None:
         filters_idx = pd.DataFrame(run_config.filters)
-        sql = sql_apply_filters_idx_to_subquery(sql, keys, filters_idx)
+        keys = [key for key in table.c if key in keys]
+        sql = sql_apply_idx_filter_to_table(sql, table, keys, filters_idx)
 
     return sql